Moves trade history types to store types
